refactor: land stranded component decomposition onto main
The event-form decomposition (EventCoverImageField #139, EventScheduleFields
#140) and the IssueDetail subscription panel (#141) were merged into
intermediate stacked branches that had already squash-merged to main, so the
changes never actually reached main. This re-applies just those component
files on top of current main (no doc/util collateral from the now-stale
branches).

- CreateEditEventFields.vue: 1045 -> 472 lines (cover-image + schedule sections
  extracted into EventCoverImageField.vue and EventScheduleFields.vue)
- IssueDetail.vue: 1009 -> 975 lines (subscription UI extracted into
  IssueSubscriptionPanel.vue)

All extracted children ship with their unit tests. tsc clean, full suite green.
