+++ components/mod/IssueDetail.vue
@@ -53,8 +53,7 @@ import {
   UNSUBSCRIBE_FROM_ISSUE,
 } from '@/graphQLData/issue/mutations';
 import NotificationComponent from '@/components/NotificationComponent.vue';
-import PrimaryButton from '@/components/PrimaryButton.vue';
-import GenericButton from '@/components/GenericButton.vue';
+import IssueSubscriptionPanel from '@/components/mod/IssueSubscriptionPanel.vue';
 import { useAutoUnsubscribe } from '@/composables/useAutoUnsubscribe';
 import { provideForumRoleMembership } from '@/composables/useForumRoleMembership';
 import { useResolvedModPermissions } from '@/composables/useResolvedModPermissions';
@@ -834,48 +833,15 @@ useAutoUnsubscribe({
     <v-row v-if="issue" class="flex justify-center dark:text-white">
       <v-col>
         <div class="px-4">
-          <div
+          <IssueSubscriptionPanel
             v-if="usernameVar && activeIssue"
-            class="mb-4 flex flex-col gap-3 rounded-lg border border-gray-200 bg-gray-50 p-4 dark:border-gray-700 dark:bg-gray-800/60"
-          >
-            <div class="flex flex-col gap-3 sm:flex-row sm:items-center sm:justify-between">
-              <div>
-                <h2 class="text-base font-semibold">Issue notifications</h2>
-                <p class="text-sm text-gray-600 dark:text-gray-300">
-                  Subscribe to replies and moderator updates on this issue.
-                </p>
-              </div>
-              <GenericButton
-                :text="isIssueSubscribed ? 'Unsubscribe' : 'Subscribe'"
-                :loading="
-                  subscribeToIssueLoading || unsubscribeFromIssueLoading
-                "
-                :active="isIssueSubscribed"
-                test-id="toggle-issue-subscription"
-                @click="toggleIssueSubscription"
-              />
-            </div>
-
-            <div
-              v-if="showSubscribeCta && !isIssueSubscribed"
-              class="rounded-lg border border-orange-200 bg-orange-50 p-3 text-sm dark:border-orange-500/40 dark:bg-orange-500/10"
-            >
-              <p class="font-medium text-gray-900 dark:text-gray-100">
-                Subscribe to updates on this issue?
-              </p>
-              <p class="mt-1 text-gray-700 dark:text-gray-300">
-                You can get notifications for replies and moderator actions.
-              </p>
-              <div class="mt-3 flex gap-2">
-                <PrimaryButton
-                  :label="'Subscribe'"
-                  :loading="subscribeToIssueLoading"
-                  @click="toggleIssueSubscription"
-                />
-                <GenericButton :text="'Not now'" @click="dismissSubscribeCta" />
-              </div>
-            </div>
-          </div>
+            :is-subscribed="isIssueSubscribed"
+            :subscribe-loading="subscribeToIssueLoading"
+            :unsubscribe-loading="unsubscribeFromIssueLoading"
+            :show-cta="showSubscribeCta"
+            @toggle="toggleIssueSubscription"
+            @dismiss-cta="dismissSubscribeCta"
+          />
 
           <h2 v-if="activeIssue" class="text-xl font-bold">Activity Feed</h2>
